requirements.txt
streamlit==1.37.1
gspread==6.1.2
google-auth==2.34.0
pandas==2.2.2
